Simplify executor by using single thread for timer driver
