+++ dds/src/std_runtime/executor.rs
@@ -1,18 +1,20 @@
 use std::{
+    collections::BinaryHeap,
     future::Future,
     pin::{Pin, pin},
     sync::{
         Arc, Mutex,
-        atomic::{self, AtomicBool},
+        atomic::{self, AtomicBool, AtomicUsize},
         mpsc::{Sender, TryRecvError, channel},
     },
     task::{Context, Poll, Wake, Waker},
     thread::{self, JoinHandle, Thread},
+    time::{Duration, Instant},
 };
 
 use crate::{
     infrastructure::error::{DdsError, DdsResult},
-    runtime::{Spawner, TaskHandle},
+    runtime::{Spawner, TaskHandle, Timer},
 };
 
 pub fn block_timeout<T>(
@@ -167,12 +169,198 @@ impl Spawner for ExecutorHandle {
     }
 }
 
-use super::timer::TimerState;
+#[derive(Debug)]
+struct TimerEntry {
+    id: usize,
+    deadline: Instant,
+    waker: Waker,
+}
+
+impl PartialEq for TimerEntry {
+    fn eq(&self, other: &Self) -> bool {
+        self.id == other.id && self.deadline == other.deadline
+    }
+}
+
+impl Eq for TimerEntry {}
+
+impl PartialOrd for TimerEntry {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for TimerEntry {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        other.deadline.cmp(&self.deadline)
+    }
+}
+
+#[derive(Default, Debug)]
+pub struct TimerState {
+    next_id: AtomicUsize,
+    entries: Mutex<BinaryHeap<TimerEntry>>,
+    executor_thread: Mutex<Option<Thread>>,
+}
+
+impl TimerState {
+    pub fn new() -> Self {
+        Self {
+            next_id: AtomicUsize::new(1),
+            entries: Mutex::new(BinaryHeap::new()),
+            executor_thread: Mutex::new(None),
+        }
+    }
+
+    pub fn set_executor_thread(&self, thread: Thread) {
+        *self.executor_thread.lock().unwrap() = Some(thread);
+    }
+
+    pub fn allocate_id(&self) -> usize {
+        self.next_id.fetch_add(1, atomic::Ordering::Relaxed)
+    }
+
+    pub fn register(&self, id: usize, deadline: Instant, waker: Waker) {
+        let mut entries = self.entries.lock().unwrap();
+        let is_earliest = entries
+            .peek()
+            .is_none_or(|earliest| deadline < earliest.deadline);
+        entries.push(TimerEntry {
+            id,
+            deadline,
+            waker,
+        });
+        drop(entries);
+
+        if is_earliest {
+            if let Some(thread) = self.executor_thread.lock().unwrap().as_ref() {
+                thread.unpark();
+            }
+        }
+    }
+
+    pub fn remove(&self, id: usize) {
+        let mut entries = self.entries.lock().unwrap();
+        entries.retain(|e| e.id != id);
+    }
+
+    pub fn duration_until_next_deadline(&self) -> Option<Duration> {
+        let entries = self.entries.lock().unwrap();
+        entries.peek().map(|earliest| {
+            let now = Instant::now();
+            if earliest.deadline > now {
+                earliest.deadline.duration_since(now)
+            } else {
+                Duration::ZERO
+            }
+        })
+    }
+
+    pub fn wake_elapsed(&self) {
+        let now = Instant::now();
+        let mut to_wake = Vec::new();
+        {
+            let mut entries = self.entries.lock().unwrap();
+            while let Some(earliest) = entries.peek() {
+                if earliest.deadline <= now {
+                    if let Some(entry) = entries.pop() {
+                        to_wake.push(entry.waker);
+                    }
+                } else {
+                    break;
+                }
+            }
+        }
+        for waker in to_wake {
+            waker.wake();
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct Sleep {
+    id: usize,
+    deadline: Option<Instant>,
+    duration: Duration,
+    timer_state: Arc<TimerState>,
+}
+
+impl Drop for Sleep {
+    fn drop(&mut self) {
+        self.timer_state.remove(self.id);
+    }
+}
+
+impl Sleep {
+    pub fn is_elapsed(&self) -> bool {
+        if let Some(d) = self.deadline {
+            Instant::now() >= d
+        } else {
+            false
+        }
+    }
+
+    pub fn reset(&mut self) {
+        self.deadline = Some(
+            Instant::now()
+                .checked_add(self.duration)
+                .unwrap_or_else(|| {
+                    Instant::now()
+                        .checked_add(Duration::from_secs(24 * 60 * 60))
+                        .unwrap_or_else(Instant::now)
+                }),
+        );
+    }
+}
+
+impl Future for Sleep {
+    type Output = ();
+
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        let this = self.get_mut();
+        if this.is_elapsed() {
+            Poll::Ready(())
+        } else {
+            if this.deadline.is_none() {
+                this.reset();
+                let deadline = this
+                    .deadline
+                    .expect("Must have deadline set after check above");
+                this.timer_state
+                    .register(this.id, deadline, cx.waker().clone());
+            }
+            Poll::Pending
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct TimerHandle {
+    state: Arc<TimerState>,
+}
+
+impl TimerHandle {
+    pub fn sleep(&self, duration: Duration) -> Sleep {
+        let id = self.state.allocate_id();
+        Sleep {
+            id,
+            deadline: None,
+            duration,
+            timer_state: self.state.clone(),
+        }
+    }
+}
+
+impl Timer for TimerHandle {
+    fn delay(&mut self, duration: core::time::Duration) -> impl Future<Output = ()> + Send {
+        self.sleep(duration)
+    }
+}
 
 pub struct Executor {
     task_sender: Sender<Arc<Task>>,
     executor_thread_handle: JoinHandle<()>,
-    timer_state_holder: Arc<Mutex<Arc<TimerState>>>,
+    timer_state: Arc<TimerState>,
 }
 
 impl Default for Executor {
@@ -183,22 +371,15 @@ impl Default for Executor {
 
 impl Executor {
     pub fn new() -> Self {
-        Self::new_with_timer_state(Arc::new(TimerState::new()))
-    }
-
-    pub fn new_with_timer_state(initial_timer_state: Arc<TimerState>) -> Self {
         let (task_sender, task_receiver) = channel::<Arc<Task>>();
-        let timer_state_holder = Arc::new(Mutex::new(initial_timer_state.clone()));
-        let timer_state_holder_clone = timer_state_holder.clone();
+        let timer_state = Arc::new(TimerState::new());
+        let timer_state_clone = timer_state.clone();
 
         let executor_thread_handle = std::thread::Builder::new()
             .name("Dust DDS Executor".to_string())
             .spawn(move || {
                 let current_thread = std::thread::current();
-                timer_state_holder_clone
-                    .lock()
-                    .unwrap()
-                    .set_executor_thread(current_thread);
+                timer_state_clone.set_executor_thread(current_thread);
 
                 loop {
                     match task_receiver.try_recv() {
@@ -222,15 +403,15 @@ impl Executor {
                             }
                         }
                         Err(TryRecvError::Empty) => {
-                            let timer_state = timer_state_holder_clone.lock().unwrap().clone();
-                            timer_state.wake_elapsed();
+                            timer_state_clone.wake_elapsed();
 
-                            if let Some(duration) = timer_state.duration_until_next_deadline() {
+                            if let Some(duration) = timer_state_clone.duration_until_next_deadline()
+                            {
                                 if duration == std::time::Duration::ZERO {
-                                    timer_state.wake_elapsed();
+                                    timer_state_clone.wake_elapsed();
                                 } else {
                                     thread::park_timeout(duration);
-                                    timer_state.wake_elapsed();
+                                    timer_state_clone.wake_elapsed();
                                 }
                             } else {
                                 thread::park();
@@ -245,20 +426,20 @@ impl Executor {
         Self {
             task_sender,
             executor_thread_handle,
-            timer_state_holder,
+            timer_state,
         }
-    }
-
-    pub fn set_timer_state(&self, timer_state: Arc<TimerState>) {
-        timer_state.set_executor_thread(self.executor_thread_handle.thread().clone());
-        *self.timer_state_holder.lock().unwrap() = timer_state;
-        self.executor_thread_handle.thread().unpark();
     }
 
     pub fn handle(&self) -> ExecutorHandle {
         ExecutorHandle {
             task_sender: self.task_sender.clone(),
             thread_handle: self.executor_thread_handle.thread().clone(),
+        }
+    }
+
+    pub fn timer_handle(&self) -> TimerHandle {
+        TimerHandle {
+            state: self.timer_state.clone(),
         }
     }
 }
