+++ dds/src/std_runtime/mod.rs
@@ -1,10 +1,8 @@
 pub mod executor;
-pub mod timer;
 
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use executor::{Executor, ExecutorHandle};
-use timer::{TimerDriver, TimerHandle};
+use executor::{Executor, ExecutorHandle, TimerHandle};
 
 use crate::{
     infrastructure::time::Time,
@@ -24,28 +22,15 @@ impl Clock for StdClock {
     }
 }
 
+#[derive(Default)]
 pub struct StdRuntime {
-    timer_driver: TimerDriver,
     executor: Executor,
 }
 
-impl Default for StdRuntime {
-    fn default() -> Self {
-        let timer_driver = TimerDriver::new();
-        let executor = Executor::new_with_timer_state(timer_driver.state());
-        Self {
-            timer_driver,
-            executor,
-        }
-    }
-}
-
 impl StdRuntime {
-    pub fn new(executor: Executor, timer_driver: TimerDriver) -> Self {
-        executor.set_timer_state(timer_driver.state());
+    pub fn new() -> Self {
         Self {
-            executor,
-            timer_driver,
+            executor: Executor::new(),
         }
     }
 }
@@ -56,7 +41,7 @@ impl DdsRuntime for StdRuntime {
     type SpawnerHandle = ExecutorHandle;
 
     fn timer(&self) -> Self::TimerHandle {
-        self.timer_driver.handle()
+        self.executor.timer_handle()
     }
 
     fn clock(&self) -> Self::ClockHandle {
